added profile, updated folder structure
