+++ frontend/public/index.html
@@ -9,11 +9,11 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="MERN Stack Authentication App - Secure login and registration system"
+      content="Movie App - Search for your favorite movies"
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>MERN Auth App</title>
+    <title>Movie App</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
