+++ frontend/src/App.tsx
@@ -7,6 +7,7 @@ import Register from "./components/Register"
 import Dashboard from "./components/Dashboard"
 import MovieDetailPage from './components/MovieDetailPage'
 import FavoritesPage from './components/FavoritesPage'
+import ProfilePage from './components/ProfilePage';
 import AppDrawer from './components/AppDrawer'
 
 import {
@@ -15,7 +16,8 @@ import {
   PATH_DASHBOARD,
   PATH_HOME,
   PATH_MOVIE_DETAIL,
-  PATH_FAVORITES
+  PATH_FAVORITES,
+  PATH_PROFILE
 } from "./constants/appConstants" 
 
 function App() {
@@ -37,6 +39,7 @@ function App() {
             >
               <Route path={PATH_DASHBOARD} element={<Dashboard />} />
               <Route path={PATH_FAVORITES} element={<FavoritesPage />} />
+              <Route path={PATH_PROFILE} element={<ProfilePage />} />
               <Route path={PATH_MOVIE_DETAIL} element={<MovieDetailPage />} />
             </Route>
             {/* Redirect root path to dashboard */}
