+++ frontend/src/components/Dashboard/index.tsx
@@ -1,8 +1,6 @@
-// src/pages/Dashboard.tsx
-import type React from "react";
-import { useState, useEffect, useCallback, useContext } from "react";
-import { useAuth } from "../context/AuthContext";
-import { PATH_LOGIN} from "../constants/appConstants";
+import React, { useState, useEffect, useCallback } from "react";
+import { useAuth } from "../../context/AuthContext";
+import { PATH_LOGIN } from "../../constants/appConstants";
 import { useNavigate } from "react-router-dom";
 import {
   Container,
@@ -14,20 +12,20 @@ import {
   Pagination,
   useTheme,
 } from "@mui/material";
-import { SxProps, Theme, darken } from "@mui/material/styles";
 
-import SearchBar from "../components/SearchBar";
-import MovieCard from "../components/MovieCard";
-import { searchMovies } from "../services/movieService";
-import { MovieSearchResult } from "../types";
-import { FE_OMDB_DEFAULT_QUERY } from "../constants/appConstants";
-import { getFavorites } from '../services/movieService';
-
+import SearchBar from "../SearchBar";
+import MovieCard from "../MovieCard";
+import { searchMovies, getFavorites } from "../../services/movieService";
+import { MovieSearchResult } from "../../types";
+import { FE_OMDB_DEFAULT_QUERY } from "../../constants/appConstants";
+import { getDashboardStyles } from './styles';
+import { MovieListParams } from './types';
 
 const Dashboard: React.FC = () => {
   const { logout, user: currentUser, isLoading: authLoading, error: authError, token } = useAuth();
   const navigate = useNavigate();
-  const theme = useTheme<Theme>();
+  const theme = useTheme();
+  const styles = getDashboardStyles(theme);
 
   const [movies, setMovies] = useState<MovieSearchResult[]>([]);
   const [totalResults, setTotalResults] = useState(0);
@@ -35,8 +33,7 @@ const Dashboard: React.FC = () => {
   const [movieFetchError, setMovieFetchError] = useState<string | null>(null);
   const [favorites, setFavorites] = useState<string[]>([]);
 
-
-  const [movieListParams, setMovieListParams] = useState({
+  const [movieListParams, setMovieListParams] = useState<MovieListParams>({
     query: FE_OMDB_DEFAULT_QUERY,
     page: 1,
   });
@@ -111,124 +108,6 @@ const Dashboard: React.FC = () => {
   );
 
   const totalPages = Math.ceil(totalResults / 10);
-
-  const styles: { [key: string]: SxProps<Theme> } = {
-    rootContainer: {
-      minHeight: '100vh',
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      py: { xs: 4, sm: 6, md: 12 },
-      px: { xs: 2, sm: 3, md: 4 },
-      background: theme.palette.gradients.darkPrimary,
-      backgroundSize: 'cover',
-      backgroundPosition: 'center',
-      overflow: 'hidden',
-      position: 'relative',
-      color: theme.palette.text.primary,
-    },
-    dashboardContent: {
-      width: '120%',
-      maxWidth: 1280,
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      gap: 4,
-      p: { xs: 2, sm: 3, md: 4 },
-      borderRadius: theme.shape.borderRadius,
-    },
-    headerBox: {
-      width: '100%',
-      display: 'flex',
-      flexDirection: { xs: 'column', sm: 'row' },
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      mb: 2,
-      gap: 2,
-    },
-    title: {
-      fontWeight: 700,
-      color: theme.palette.primary.light,
-      textShadow: `0 0 10px ${theme.palette.glow.main}60`,
-    },
-    logoutButton: {
-      py: 1,
-      px: 3,
-      borderRadius: theme.shape.borderRadius,
-      bgcolor: theme.palette.error.main,
-      color: theme.palette.error.contrastText,
-      boxShadow: `0 4px 8px ${darken(theme.palette.error.main, 0.3)}`,
-      transition: 'background-color 0.3s ease-in-out, box-shadow 0.3s ease-in-out, transform 0.2s ease-in-out',
-      '&:hover': {
-        bgcolor: theme.palette.error.dark,
-        boxShadow: `0 6px 12px ${darken(theme.palette.error.main, 0.5)}`,
-        transform: 'translateY(-2px)',
-      },
-      '&:active': {
-        transform: 'translateY(0)',
-        boxShadow: `0 2px 8px ${darken(theme.palette.error.main, 0.3)}`,
-      },
-    },
-    searchSection: {
-      width: '100%',
-      display: 'flex',
-      justifyContent: 'center',
-      mb: 4,
-    },
-    resultsSection: {
-      width: '100%',
-      minHeight: '200px',
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    loadingIndicator: {
-      mt: 4,
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      gap: 2,
-      color: theme.palette.text.primary,
-    },
-    errorAlert: {
-      mb: 2,
-      borderRadius: theme.shape.borderRadius,
-      width: '100%',
-      boxShadow: theme.shadows[2],
-    },
-    noResults: {
-      color: theme.palette.text.secondary,
-      mt: 4,
-      textAlign: 'center',
-      lineHeight: 1.5,
-    },
-    movieGrid: {
-      display: 'grid',
-      gridTemplateColumns: 'repeat(auto-fill, minmax(200px, 1fr))',
-      justifyContent: 'center',
-      gap: theme.spacing(9),
-      width: '100%',
-      alignItems: 'center',
-      justifyItems: 'center', // Center items in their grid cells
-      '@media (min-width: 600px)': {
-        gridTemplateColumns: 'repeat(auto-fill, minmax(220px, 1fr))',
-      },
-      '@media (min-width: 960px)': {
-        gridTemplateColumns: 'repeat(auto-fill, minmax(250px, 1fr))',
-      },
-    },
-    movieCardWrapper: {
-      display: 'flex',
-      justifyContent: 'center',
-    },
-    paginationContainer: {
-      mt: 4,
-      mb: 2,
-      display: 'flex',
-      justifyContent: 'center',
-    }
-  };
 
   return (
     <Box sx={styles.rootContainer}>
@@ -311,4 +190,4 @@ const Dashboard: React.FC = () => {
   );
 };
 
-export default Dashboard;
+export default Dashboard;
