+++ frontend/src/components/FavoritesPage/index.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { Box, Typography, CircularProgress, Alert, useTheme, Container } from "@mui/material";
-import { useAuth } from "../context/AuthContext";
-import { getFavoriteMoviesDetails, removeFavorite } from "../services/movieService";
-import MovieCard from "./MovieCard";
-import { MovieSearchResult } from "../types";
+import { useAuth } from "../../context/AuthContext";
+import { getFavoriteMoviesDetails, removeFavorite } from "../../services/movieService";
+import MovieCard from "../MovieCard";
+import { MovieSearchResult } from "../../types";
 import { useNavigate } from "react-router-dom";
+import { getFavoritesPageStyles } from './styles';
 
 const FavoritesPage: React.FC = () => {
   const { token } = useAuth();
@@ -13,6 +14,7 @@ const FavoritesPage: React.FC = () => {
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const navigate = useNavigate();
+  const styles = getFavoritesPageStyles(theme);
 
   useEffect(() => {
     const fetchFavorites = async () => {
@@ -38,51 +40,6 @@ const FavoritesPage: React.FC = () => {
     if (!res.error) {
       setMovies((prev) => prev.filter((m) => m.imdbID !== imdbID));
     }
-  };
-
-  const styles = {
-    root: {
-      minHeight: "100vh",
-      display: "flex",
-      flexDirection: "column",
-      alignItems: "center",
-      py: { xs: 4, sm: 6, md: 12 },
-      px: { xs: 2, sm: 3, md: 4 },
-      background: theme.palette.gradients.darkPrimary,
-      backgroundSize: "cover",
-      backgroundPosition: "center",
-      overflow: "hidden",
-      position: "relative",
-      color: theme.palette.text.primary,
-    },
-    movieGrid: {
-      display: "grid",
-      gridTemplateColumns: "repeat(4, 1fr)",
-      justifyContent: "center",
-      gap: theme.spacing(9),
-      width: "100%",
-      alignItems: "center",
-      justifyItems: "center",
-      "@media (max-width: 1200px)": {
-        gridTemplateColumns: "repeat(3, 1fr)",
-      },
-      "@media (max-width: 900px)": {
-        gridTemplateColumns: "repeat(2, 1fr)",
-      },
-      "@media (max-width: 600px)": {
-        gridTemplateColumns: "repeat(1, 1fr)",
-      },
-    },
-    movieCardWrapper: {
-      display: "flex",
-      justifyContent: "center",
-    },
-    noResults: {
-      color: theme.palette.text.secondary,
-      mt: 4,
-      textAlign: "center",
-      lineHeight: 1.5,
-    },
   };
 
   const handleCardClick = (imdbID: string) => {
