+++ frontend/src/index.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
-import theme from './theme'; // Import your custom theme
+import theme from './theme';
 import "./index.css"
 import App from "./App"
