+++ frontend/src/types/index.ts
@@ -2,6 +2,7 @@ export interface User {
   id: string
   name: string
   email: string
+  createdAt?: string;
 }
 
 export interface AuthState {
